split up generic infer/validate
