+++ ark/type/parser/shift/operand/unenclosed.ts
@@ -21,7 +21,8 @@ import {
 } from "@ark/util"
 import type { ArkAmbient } from "../../../config.ts"
 import type { resolutionToAst } from "../../../scope.ts"
-import type { GenericInstantiationAst, InferredAst } from "../../ast/infer.ts"
+import type { GenericInstantiationAst } from "../../ast/generic.ts"
+import type { InferredAst } from "../../ast/infer.ts"
 import { writePrefixedPrivateReferenceMessage } from "../../ast/validate.ts"
 import type { DynamicState } from "../../reduce/dynamic.ts"
 import type { StaticState, state } from "../../reduce/static.ts"
